+++ src/procedure_geometry.h
@@ -9,5 +9,5 @@ class LineMesh;
 
 void create_floor(std::vector<glm::vec4>& floor_vertices, std::vector<glm::uvec3>& floor_faces);
 // FIXME: Add functions to generate the bone mesh.
-void create_skel(Mesh& mesh, std::vector<glm::vec4>& skel_vertices, std::vector<glm::vec2>& skel_lines);
+void create_skel(Mesh& mesh, std::vector<glm::vec4>& skel_vertices, std::vector<glm::uvec2>& skel_lines);
 #endif
